backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0

# Environment and configuration
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Scientific computing
numpy==1.26.2
scipy==1.11.4

# Deep learning (PyTorch)
torch==2.1.1
torchvision==0.16.1

# 3D processing
trimesh==4.0.5
open3d==0.18.0
pyglet<2

# Depth estimation models
timm==0.9.12
transformers==4.36.0

# File handling
python-magic==0.4.27
werkzeug==3.0.1

# Async task processing (future)
celery==5.3.4
redis==5.0.1

# Database drivers
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code quality
black==23.12.0
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Monitoring (optional)
sentry-sdk[flask]==1.38.0

# API documentation
flask-swagger-ui==4.11.1
